docs/archive/freqtrade_qlib_engineering_draft/requirements.txt
# requirements.txt
# Freqtrade + Qlib 项目依赖

# 核心框架
freqtrade>=2024.9
qlib>=0.9.0

# 因子分析
alphalens-reloaded>=0.4.4

# 数据处理
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0
pytz>=2023.3

# 机器学习
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
catboost>=1.2.0

# 时间序列和数据获取
CCXT>=4.0.0
yfinance>=0.2.0

# Jupyter 和可视化
jupyter>=1.0.0
jupyterlab>=3.5.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# 配置和工具
pyyaml>=6.0
python-dotenv>=1.0.0
requests>=2.31.0
joblib>=1.3.0

# 可选：高级功能
# ta>=0.10.0                 # 技术指标库
# hmmlearn>=0.3.0            # 隐马尔可夫模型
# scikit-optimize>=0.9.0     # 贝叶斯优化
